+++ NEWS.md
@@ -1,3 +1,15 @@
+# MyPackage 0.1.3
+
+## New Features
+- Change name package to caIRA.
+
+## Improvements
+- Loading bar.
+
+## Bug Fixes
+- Making function faster than before.
+
+
 # MyPackage 0.1.2
 
 ## New Features